Move to vector internally.
Earlier use of set works, but appeared to be much faster than it
actually was. std::set::merge was deleting the contents of the set
being merged into the pattern set. As a result, after a while all
of the pattern sets would be deleted and there would be no patterns
to match against == looks like good performance.